A bit of refactgoring
