+++ .gitignore
@@ -2,3 +2,4 @@ python
 *~
 .mypy_cache/
 .idea/
+__pycache__/
